+++ config/application_test.go
@@ -4,6 +4,7 @@ import (
 	"github.com/dogmatiq/dogma"
 	. "github.com/dogmatiq/enginekit/config"
 	"github.com/dogmatiq/enginekit/fixtures"
+	"github.com/dogmatiq/enginekit/identity"
 	"github.com/dogmatiq/enginekit/message"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
@@ -114,15 +115,11 @@ var _ = Describe("type ApplicationConfig", func() {
 				))
 			})
 
-			Describe("func Name()", func() {
-				It("returns the app name", func() {
-					Expect(cfg.Name()).To(Equal("<app>"))
-				})
-			})
-
-			Describe("func Key()", func() {
-				It("returns the app key", func() {
-					Expect(cfg.Key()).To(Equal("<app-key>"))
+			Describe("func Identity()", func() {
+				It("returns the app identity", func() {
+					Expect(cfg.Identity()).To(Equal(
+						identity.MustNew("<app>", "<app-key>"),
+					))
 				})
 			})
 		})
@@ -174,7 +171,7 @@ var _ = Describe("type ApplicationConfig", func() {
 
 				Expect(err).To(Equal(
 					Error(
-						`*fixtures.Application.Configure() called ApplicationConfigurer.Identity() with an invalid name "\t \n"`,
+						`*fixtures.Application.Configure() called ApplicationConfigurer.Identity() with an invalid name "\t \n", names must be non-empty, printable UTF-8 strings with no whitespace`,
 					),
 				))
 			})
@@ -192,7 +189,7 @@ var _ = Describe("type ApplicationConfig", func() {
 
 				Expect(err).To(Equal(
 					Error(
-						`*fixtures.Application.Configure() called ApplicationConfigurer.Identity() with an invalid key "\t \n"`,
+						`*fixtures.Application.Configure() called ApplicationConfigurer.Identity() with an invalid key "\t \n", keys must be non-empty, printable UTF-8 strings with no whitespace`,
 					),
 				))
 			})