+++ config/config.go
@@ -5,18 +5,15 @@ import (
 	"reflect"
 
 	"github.com/dogmatiq/enginekit/handler"
+	"github.com/dogmatiq/enginekit/identity"
 	"github.com/dogmatiq/enginekit/message"
 )
 
 // Config is an interface for all configuration values.
 type Config interface {
-	// Name returns the name of the configured item.
-	// For example, the application or handler name.
-	Name() string
-
-	// Key returns the key of the configured item.
-	// For example, the application or handler unique key.
-	Key() string
+	// Identity returns the identity of the configured item.
+	// For example, the application or handler identity.
+	Identity() identity.Identity
 
 	// Accept calls the appropriate method on v for this configuration type.
 	Accept(ctx context.Context, v Visitor) error