+++ handler/errors_test.go
@@ -3,6 +3,7 @@ package handler_test
 import (
 	"github.com/dogmatiq/enginekit/fixtures"
 	. "github.com/dogmatiq/enginekit/handler"
+	"github.com/dogmatiq/enginekit/identity"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 )
@@ -11,8 +12,7 @@ var _ = Describe("type EmptyInstanceIDError", func() {
 	Describe("func Error", func() {
 		It("returns a meaningful error description", func() {
 			err := EmptyInstanceIDError{
-				HandlerName: "<name>",
-				HandlerKey:  "<key>",
+				Handler:     identity.MustNew("<name>", "<key>"),
 				HandlerType: AggregateType,
 				Message:     fixtures.MessageA1,
 			}
@@ -28,8 +28,7 @@ var _ = Describe("type NilRootError", func() {
 	Describe("func Error", func() {
 		It("returns a meaningful error description", func() {
 			err := NilRootError{
-				HandlerName: "<name>",
-				HandlerKey:  "<key>",
+				Handler:     identity.MustNew("<name>", "<key>"),
 				HandlerType: AggregateType,
 			}
 
@@ -45,8 +44,7 @@ var _ = Describe("type EventNotRecordedError", func() {
 		When("the instance was created", func() {
 			It("returns a meaningful error description", func() {
 				err := EventNotRecordedError{
-					HandlerName:  "<name>",
-					HandlerKey:   "<key>",
+					Handler:      identity.MustNew("<name>", "<key>"),
 					WasDestroyed: false,
 					Message:      fixtures.MessageA1,
 					InstanceID:   "<instance>",
@@ -61,8 +59,7 @@ var _ = Describe("type EventNotRecordedError", func() {
 		When("the instance was destroyed", func() {
 			It("returns a meaningful error description", func() {
 				err := EventNotRecordedError{
-					HandlerName:  "<name>",
-					HandlerKey:   "<key>",
+					Handler:      identity.MustNew("<name>", "<key>"),
 					WasDestroyed: true,
 					Message:      fixtures.MessageA1,
 					InstanceID:   "<instance>",
@@ -80,8 +77,7 @@ var _ = Describe("type UnexpectedMessageError", func() {
 	Describe("func Error", func() {
 		It("returns a meaningful error description", func() {
 			err := UnexpectedMessageError{
-				HandlerName: "<name>",
-				HandlerKey:  "<key>",
+				Handler:     identity.MustNew("<name>", "<key>"),
 				HandlerType: AggregateType,
 				Message:     fixtures.MessageA1,
 				InstanceID:  "<instance>",